DSPLLE: sort of semi-working breakpoints and stepping, if you flip an #ifdef. more work to do, for some reason it gets very slow when you enable it atm
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3573 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DSPCore/Src/DSPStacks.cpp
@@ -30,27 +30,27 @@
 
 // Stacks. The stacks are outside the DSP RAM, in dedicated hardware.
 
-void dsp_reg_stack_push(u8 stack_reg)
+void dsp_reg_stack_push(int stack_reg)
 {
 	g_dsp.reg_stack_ptr[stack_reg]++;
 	g_dsp.reg_stack_ptr[stack_reg] &= DSP_STACK_MASK;
 	g_dsp.reg_stack[stack_reg][g_dsp.reg_stack_ptr[stack_reg]] = g_dsp.r[DSP_REG_ST0 + stack_reg];
 }
 
-void dsp_reg_stack_pop(u8 stack_reg)
+void dsp_reg_stack_pop(int stack_reg)
 {
 	g_dsp.r[DSP_REG_ST0 + stack_reg] = g_dsp.reg_stack[stack_reg][g_dsp.reg_stack_ptr[stack_reg]];
 	g_dsp.reg_stack_ptr[stack_reg]--;
 	g_dsp.reg_stack_ptr[stack_reg] &= DSP_STACK_MASK;
 }
 
-void dsp_reg_store_stack(u8 stack_reg, u16 val)
+void dsp_reg_store_stack(int stack_reg, u16 val)
 {
 	dsp_reg_stack_push(stack_reg);
 	g_dsp.r[DSP_REG_ST0 + stack_reg] = val;
 }
 
-u16 dsp_reg_load_stack(u8 stack_reg)
+u16 dsp_reg_load_stack(int stack_reg)
 {
 	u16 val = g_dsp.r[DSP_REG_ST0 + stack_reg];
 	dsp_reg_stack_pop(stack_reg);
