+++ Source/Core/DSPCore/Src/DSPStacks.h
@@ -28,7 +28,7 @@
 
 #include "Common.h"
 
-void dsp_reg_store_stack(u8 stack_reg, u16 val);
-u16 dsp_reg_load_stack(u8 stack_reg);
+void dsp_reg_store_stack(int stack_reg, u16 val);
+u16 dsp_reg_load_stack(int stack_reg);
 
 #endif
